Revet docker config
